+++ src/powernse/reading/__init__.py
@@ -7,6 +7,8 @@ from powernse.reading.deals import DealsReader, SecbanReader
 from powernse.reading.delivery import DeliveryReader
 from powernse.reading.futures import FoReader
 from powernse.reading.index import IndexReader
+from powernse.reading.manifest_audit import ManifestAudit, ManifestIssue
+from powernse.reading.security import SecurityMasterReader
 from powernse.reading.snapshots import SnapshotReader
 
 __all__ = [
@@ -17,6 +19,9 @@ __all__ = [
     "DeliveryReader",
     "FoReader",
     "IndexReader",
+    "ManifestAudit",
+    "ManifestIssue",
     "SecbanReader",
+    "SecurityMasterReader",
     "SnapshotReader",
 ]